refactor: rename build.js and relocate parse-utils.js for better organization
1. Renamed build.js → index-status.js
   - File only contains getIndexStatus() and checkReindexEligibility()
   - New name accurately reflects its focused purpose

2. Moved core/parse-utils.js → indexing/parse-utils.js
   - All 7 consumers are in indexing/ directory
   - Utilities are indexing-specific, not general-purpose
   - Maintained backward compatibility via re-exports in core/utils.js and core/paths.js

Updated all import paths across codebase and tests.
