+++ nx/blocks/media-library/indexing/medialog.js
@@ -6,7 +6,7 @@ import {
   detectMediaType,
   computeCanonicalMetadata,
 } from './parse.js';
-import { normalizePath } from '../core/parse-utils.js';
+import { normalizePath } from './parse-utils.js';
 import { canonicalizeMediaUrl } from '../core/urls.js';
 
 /**
